feat(core): aggregate store
Definition of ReadModel and implementation of an aggregate store which
maintains a cache and responds to query with the latest version.
If the version in cache is older than a given lag, new events are fetched
from the stream and applied to the current version.

It is still unusable because there is no exported function to create
such an aggregate store.