feat(postgres): stream family migrations
Database.CQRS.PostgreSQL.Migration.migrate is feeding all events from a
stream family to a transformer to populate a new table and swap it with
the current one.

It can be used to migrate events of a stream family by dropping, merging
or re-encoding events while archiving the old streams.