feat(postgres): allow execution of custom actions in projections
`runProjectionWith` now takes a consumer to commit the actions and
update the tracking table.

For SQL actions, using `executeSQLActions` yields the old behaviour.
There is also `executeCustomActions` which takes a function of type
`action -> m (Either String (m ()))` to run each action returning either
an error or a rollback action.