feat(core): task manager
A new module `Database.CQRS.TaskManager` provides a way to run tasks
from a simple aggregator on a stream family.

The aggregator must be of type `event -> St.State ([Task action], st) ()`.

Tasks can be "planned", i.e. they have a `T.UTCTime` from which they
should be run.

The task manager is looping through streams in the family and execute the
tasks it finds. If it's left without tasks to run, it waits 5 minutes or
less if it knows of an earlier planned task.