fix(postgres): stop listening thread from quitting too early
The notification listening thread (`allNewEvents`) terminates when the
queue has been garbage-collected. It uses a weak pointer to the queue
and checks that the queue is still alive after receiving a notification.

The weak pointer's key was dying prematurely probably due to some
optimisation from GHC. It now uses another one which doesn't seem to
cause the same issue:

data GCKey = GCKey