fix(core): stricter interpretation of lag tolerance
To use the cached value in AggregateStore without looking for new events, the
current time needs to be *strictly* lower than the previous time + lag tolerance.

If the lag tolerance is set to 0, it should always check for new events. I suspect
that if, say, the task manager's loop is too fast, the current time will not change
because of the limited precision of the clock so using <= doesn't work in that
case.