feat: refactoring of tabular data action
There are now two ways of defining columns: `Flat` is just a list of
columns while `WithUniqueKey` allows to separate the columns used as a
unique key from the rest of the columns.

This allows the library to define an in-memory `Table` for these types.
Tuples are stored in a list or a hash map (using the key) respectively.

Optimisations have been moved to
`Database.CQRS.TabularData.Optimisation` and are not used by
`eventsourcing-postgresql` by default anymore. There is no guarantee
that it would be any faster after all.