feat(core): refactoring of projections
`Update` now takes a tuple of type `Tuple Update cols` (using `Update`
instead of `Last`.) At the moment, `Update` is very basic and it allows
to set the column to a specific value or to increment or decrement it by
a given value.

When a projection fails on a batch of events, it splits it in half
(recursively) and retries each batch in turn. This ensures that
projections and aggregators agree on the current state of an aggregate
and that projections run on as many events as possible.

The projection and tabular data logic has tests.