feat(core): interruptible stream family
Add a new `InterruptibleStreamFamily` which wraps a stream family together
with a function to determine whether the family should stop producing
new events.

For example, `timingOutStreamFamily` stops producing new events when no
new events have been received in a given period of time.

This can be used to bring down a projection if it's inactive instead of
constantly running it in the background.