feat: top up read models
Given an underlying read model, a stream family and an aggregator, we
can build a new read model that fetches the last projected state, any
new events that were added since the last run of the projection and runs
the aggregator on them.

This means there can be a projection running in the background writing
to a database with a read model that always returns the aggregate based
on all events regardless of how fast the projection is.